refactor: de-indy-ify AnonCredsRevocation
